Put workspace reducer in utils file
